完成员工添加并修复部门管理几个问题

+++ src/components/sys/basic/DepMana.vue
@@ -99,6 +99,9 @@
                     let d = deps[i];
                     if (d.id == dep.parentId) {
                         d.children = d.children.concat(dep);
+                        if (d.children.length > 0) {
+                            d.isParent = true;
+                        }
                         return;
                     } else {
                         this.addDep2Deps(d.children, dep);
@@ -120,34 +123,41 @@
                 this.dep.parentId = data.id;
                 this.dialogVisible = true;
             },
-            removeDepFromDeps(deps, id) {
+            removeDepFromDeps(p, deps, id) {
                 for (let i = 0; i < deps.length; i++) {
                     let d = deps[i];
                     if (d.id == id) {
                         deps.splice(i, 1);
+                        if (deps.length == 0) {
+                            p.isParent = false;
+                        }
                         return;
                     } else {
-                        this.removeDepFromDeps(d.children, id);
+                        this.removeDepFromDeps(d, d.children, id);
                     }
                 }
             },
             deleteDep(data) {
-                this.$confirm('此操作将永久删除【' + data.name + '】部门，是否继续?', '提示', {
-                    confirmButtonText: '确定',
-                    cancelButtonText: '取消',
-                    type: 'warning'
-                }).then(() => {
-                    this.deleteRequest("/system/basic/department/" + data.id).then(resp => {
-                        if (resp) {
-                            this.removeDepFromDeps(this.deps, data.id);
-                        }
-                    })
-                }).catch(() => {
-                    this.$message({
-                        type: 'info',
-                        message: '已取消删除'
+                if (data.isParent) {
+                    this.$message.error("父部门删除失败");
+                } else {
+                    this.$confirm('此操作将永久删除【' + data.name + '】部门，是否继续?', '提示', {
+                        confirmButtonText: '确定',
+                        cancelButtonText: '取消',
+                        type: 'warning'
+                    }).then(() => {
+                        this.deleteRequest("/system/basic/department/" + data.id).then(resp => {
+                            if (resp) {
+                                this.removeDepFromDeps(null, this.deps, data.id);
+                            }
+                        })
+                    }).catch(() => {
+                        this.$message({
+                            type: 'info',
+                            message: '已取消删除'
+                        });
                     });
-                });
+                }
             },
             initDeps() {
                 this.getRequest("/system/basic/department/").then(resp => {
